requirements.txt
numpy
torch
trimesh
scikit-learn
PyMCubes